Rename MembersSection to TeamSection
