--- components/Sections/Members/Content.tsx
@@ -1,25 +1,0 @@
-import React from "react";
-
-import MembersSlider from "components/Slider/MembersSlider";
-import Section from "components/Sections/Section";
-
-import { MembersSectionContentProps } from "./types";
-import { membersSectionContent, membersSectionId } from "./constants";
-
-const MembersSectionContent: React.FC<MembersSectionContentProps> = ({
-  text = membersSectionContent.text,
-  title = membersSectionContent.title,
-  members = membersSectionContent.members,
-}) => (
-  <Section
-    id={membersSectionId}
-    title={title}
-    text={text}
-  >
-    <div className="w-full mt-12">
-      <MembersSlider members={members} />
-    </div>
-  </Section>
-);
-
-export default MembersSectionContent;
