--- components/Sections/Members/index.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-import MembersSectionContent from "./Content";
-
-const MembersSection: React.FC = () => (
-  <MembersSectionContent />
-);
-
-export default MembersSection;
