+++ components/Sections/Team/index.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import TeamSectionContent from "./Content";
+
+const TeamSection: React.FC = () => (
+  <TeamSectionContent />
+);
+
+export default TeamSection;
