+++ components/Slider/MembersSlider/types.ts
@@ -1,4 +1,4 @@
-import { Member } from "components/Sections/Members/types";
+import { Member } from "components/Sections/Team/types";
 
 export interface MembersSliderProps {
   members: Member[];
