+++ pages/index.tsx
@@ -5,7 +5,7 @@ import MissionSection from "components/Sections/Mission";
 import AchievementsSection from "components/Sections/Achievements";
 import DonationsSection from "components/Sections/Donations";
 import ActionsSection from "components/Sections/Actions";
-import MembersSection from "components/Sections/Members";
+import TeamSection from "components/Sections/Team";
 
 const App: React.FC = () => (
   <>
@@ -18,7 +18,7 @@ const App: React.FC = () => (
       <ActionsSection />
       <DonationsSection />
       <MissionSection />
-      <MembersSection />
+      <TeamSection />
     </main>
   </>
 );
